feat(codeeditor): tree restructuring - Adding codeEditor structure from ui lib - Restructuring src to be more comprehensive
