+++ package.json
@@ -7,6 +7,7 @@
     "@readme/variable": "^7.1.0",
     "codemirror": "^5.48.2",
     "react": "^16.13.1",
+    "react-codemirror2": "^7.2.1",
     "react-dom": "^16.13.1"
   },
   "scripts": {
