+++ public/index.js
@@ -1,7 +1,15 @@
+import React from 'react';
 import ReactDOM from 'react-dom';
-import syntaxHighlighter from '../src/index';
+import syntaxHighlighter, { CodeEditor } from '../src/index';
 
 ReactDOM.render(
-  syntaxHighlighter('console.log("Hello, world!");', 'js', { dark: true }),
+  <div>
+    <h1>Core Syntax Highlighter</h1>
+    {syntaxHighlighter('console.log("Hello, world!");', 'js', { dark: true })}
+    <hr />
+    <h1>Code Editor</h1>
+    <CodeEditor />
+  </div>
+  ,
   document.getElementById('root')
 );
