+++ src/utils/canonical.js
